Move Code to Modules
* Move cli to a module
* Move output to a module
* Move request/http to a module
